LayoutTests/fast/css-grid-layout/grid-shorthand-get-set-expected.txt

This test checks that the 'grid' shorthand is properly parsed and the longhand properties correctly assigned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting the longhand values when shorthand is set through CSS.
PASS window.getComputedStyle(gridWithDense, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithDense, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithDense, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithDense, '').getPropertyValue('-webkit-grid-auto-flow') is "row dense"
PASS window.getComputedStyle(gridWithDense, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithDense, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('-webkit-grid-template-columns') is "15px"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('-webkit-grid-template-rows') is "10px"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTemplate1, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithTemplate1, '').getPropertyValue('-webkit-grid-template-rows') is "10px"
PASS window.getComputedStyle(gridWithTemplate1, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTemplate1, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTemplate1, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTemplate1, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTemplateNone, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithTemplateNone, '').getPropertyValue('-webkit-grid-template-rows') is "10px"
PASS window.getComputedStyle(gridWithTemplateNone, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTemplateNone, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTemplateNone, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTemplateNone, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowDenseAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-flow') is "row dense"
PASS window.getComputedStyle(gridWithAutoFlowDenseAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowDenseAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-rows') is "20px"
PASS window.getComputedStyle(gridWithAutoFlowAndColumns, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndColumns, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndColumns, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndColumns, '').getPropertyValue('-webkit-grid-auto-flow') is "column"
PASS window.getComputedStyle(gridWithAutoFlowAndColumns, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowAndColumns, '').getPropertyValue('-webkit-grid-auto-rows') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowDense, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDense, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDense, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDense, '').getPropertyValue('-webkit-grid-auto-flow') is "row dense"
PASS window.getComputedStyle(gridWithAutoFlowDense, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowDense, '').getPropertyValue('-webkit-grid-auto-rows') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-auto-flow') is "column dense"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('-webkit-grid-auto-rows') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-auto-flow') is "row dense"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('-webkit-grid-auto-rows') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-flow') is "column"
PASS window.getComputedStyle(gridWithAutoFlowAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-rows') is "20px"

Test setting wrong values for 'grid' shorthand through CSS (they should resolve to the default: 'row')
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowNoneAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowNoneAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowNoneAndColumnsAndRows, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowNoneAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowNoneAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowNoneAndColumnsAndRows, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowNone, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowNone, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowNone, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowNone, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowNone, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowNone, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExplicitAndImplicit, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExplicitAndImplicit, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExplicitAndImplicit, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExplicitAndImplicit, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExplicitAndImplicit, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExplicitAndImplicit, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithWrongSlash1, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithWrongSlash1, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithWrongSlash1, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithWrongSlash1, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithWrongSlash1, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithWrongSlash1, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithWrongSlash2, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithWrongSlash2, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithWrongSlash2, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithWrongSlash2, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithWrongSlash2, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithWrongSlash2, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithWrongSlash3, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithWrongSlash3, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithWrongSlash3, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithWrongSlash3, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithWrongSlash3, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithWrongSlash3, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowRowAndColumn, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowRowAndColumn, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowRowAndColumn, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowRowAndColumn, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowRowAndColumn, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowRowAndColumn, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithoutAutoFlowAndExtraBreath, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithoutAutoFlowAndExtraBreath, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithoutAutoFlowAndExtraBreath, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithoutAutoFlowAndExtraBreath, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithoutAutoFlowAndExtraBreath, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithoutAutoFlowAndExtraBreath, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowString1, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString1, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString1, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString1, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowString1, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowString1, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowString2, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString2, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString2, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString2, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowString2, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowString2, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowString3, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString3, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString3, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowString3, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowString3, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowString3, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTemplateAndAutoFlow, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithTemplateAndAutoFlow, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithTemplateAndAutoFlow, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTemplateAndAutoFlow, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTemplateAndAutoFlow, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTemplateAndAutoFlow, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTemplateAndMisplacedString1, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithTemplateAndMisplacedString1, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithTemplateAndMisplacedString1, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTemplateAndMisplacedString1, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTemplateAndMisplacedString1, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTemplateAndMisplacedString1, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"

Test getting and setting 'grid' shorthand through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "10px"
PASS element.style.webkitGridTemplateColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "20px"
PASS element.style.webkitGridTemplateRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS element.style.webkitGridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS element.style.webkitGridAutoFlow is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS element.style.webkitGridAutoColumns is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS element.style.webkitGridAutoRows is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "10px"
PASS element.style.webkitGridTemplateColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "[line] 20px"
PASS element.style.webkitGridTemplateRows is "[line] 20px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "\"a\""
PASS element.style.webkitGridTemplateAreas is "\"a\""
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS element.style.webkitGridAutoFlow is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS element.style.webkitGridAutoColumns is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS element.style.webkitGridAutoRows is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS element.style.webkitGridTemplateColumns is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS element.style.webkitGridTemplateRows is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS element.style.webkitGridTemplateAreas is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "row dense"
PASS element.style.webkitGridAutoFlow is "row dense"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "20px"
PASS element.style.webkitGridAutoColumns is "20px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "20px"
PASS element.style.webkitGridAutoRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS element.style.webkitGridTemplateColumns is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS element.style.webkitGridTemplateRows is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS element.style.webkitGridTemplateAreas is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "column"
PASS element.style.webkitGridAutoFlow is "column"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "20px"
PASS element.style.webkitGridAutoColumns is "20px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "10px"
PASS element.style.webkitGridAutoRows is "10px"

Test the initial value
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is 'row'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'

Test setting grid-template-columns and grid-template-rows back to 'none' through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS element.style.webkitGridTemplateColumns is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS element.style.webkitGridTemplateRows is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS element.style.webkitGridTemplateAreas is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "column"
PASS element.style.webkitGridAutoFlow is "column"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "10px"
PASS element.style.webkitGridAutoColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "20px"
PASS element.style.webkitGridAutoRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS element.style.webkitGridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS element.style.webkitGridTemplateRows is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-areas') is "none"
PASS element.style.webkitGridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-flow') is "row"
PASS element.style.webkitGridAutoFlow is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is "auto"
PASS element.style.webkitGridAutoColumns is "initial"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is "auto"
PASS element.style.webkitGridAutoRows is "initial"

Test the inherit value on reset-only subproperties (grid-*-gap)
PASS getComputedStyle(anotherElement, '').getPropertyValue('-webkit-grid-column-gap') is "0px"
PASS getComputedStyle(anotherElement, '').getPropertyValue('-webkit-grid-row-gap') is "0px"
PASS getComputedStyle(anotherElement, '').getPropertyValue('-webkit-grid-column-gap') is "20px"
PASS getComputedStyle(anotherElement, '').getPropertyValue('-webkit-grid-row-gap') is "100px"
PASS successfullyParsed is true

TEST COMPLETE
